Move LoginPage and RegistrationPage to containers
